Refactored editor assets to be in more correct locations
